backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
requests==2.31.0
beautifulsoup4==4.12.2
#spacy==3.5.1
scikit-learn==1.2.2
pandas==2.0.2
pytest==7.4.0
PyJWT==2.8.0
Werkzeug==2.2.3
python-dotenv==1.0.0
nltk==3.8.1
Flask-SocketIO==5.3.5
eventlet==0.40.3
